Add a second dashboard client
This client is for dashboard-staging, dashboard-demo and local dev

Long term, when creds are rotated, the creds will be pushed to the UPS
in the space of the app.

While it's easy to track the creds for the deployed apps in the UPS, it
would be complicated to maintain a separate env for local dev because
no app would be bound to it since it's local.
Now with the local dev as the same staging one, a dev could just read
the env of the staging app and get the creds.