Allow generating missing certificate
In cf v24.6.0, the router job now expects this certificate to exist in the secret store. We stopped generating it when we added these lines to the opsfile in cfb7e89. The secrets remained in the secret store in dev and staging, but were removed in production. Now that they're required by the job, dev and staging are using the previously-generated certs that remain in the secret store, but prod is failing because the secret was deleted. This change should allow the cert to be regenerated.