Declare scopes for billing UAA client
We are adding an admin API to the billing service. We will add Cloud.gov operators to the `usage.admin` group in UAA and the billing service will enforce authorization using the scope. `usage.read` is intended for the user-facing API, which is not yet implemented.

Related to https://github.com/cloud-gov/product/issues/3303.